what2eat/requirements.txt
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
certifi==2020.12.5
colorama==0.4.3
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.14
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
